archive/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "All Things Linux Infrastructure Management"
readme = "README.md"
requires-python = ">=3.13.2"
authors = [{ name = "All Things Linux", email = "admin@allthingslinux.com" }]

dependencies = [
    "ansible-core>=2.18.6",
    "ansible-runner>=2.4.1",
    "click>=8.2.1",
    "docker>=7.1.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-terraform>=0.10.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "tabulate>=0.9.0",
    "urllib3>=2.5.0",
]

[project.scripts]
atl = "scripts.cli:cli"

[project.urls]
Repository = "https://github.com/allthingslinux/infra"

[build-system]
requires = ["uv_build>=0.7.19,<0.8"]
build-backend = "uv_build"

# Configure uv build backend for our project structure
[tool.uv.build-backend]
module-name = "scripts"
module-root = ""

[tool.uv.sources]
tflint-py = { git = "https://github.com/AleksaC/tflint-py.git", rev = "v0.58.0" }

[dependency-groups]
dev = [
    "ansible-dev-tools>=25.5.2",
    "ansible-lint>=25.6.1",
    "basedpyright>=1.26.2",
    "commitlint>=1.3.0",
    "lefthook>=1.11.0",
    "pymarkdownlnt>=0.9.30",
    "ruff>=0.12.2",
    "shellcheck-py>=0.10.0.1",
    "shfmt-py>=3.12.0.2",
    "tflint-py>=0.58.0",
    "yamllint>=1.37.1",
]
docs = [
    "cairosvg>=2.8.2",
    "mkdocs>=1.6.1",
    "mkdocs-include-markdown-plugin>=7.1.6",
    "mkdocs-material>=9.6.15",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "mkdocs-minify-plugin>=0.8.0",
    "pillow>=11.3.0",
    "pymdown-extensions>=10.16",
]
test = [
    "molecule>=25.6.0",
    "molecule-plugins>=23.7.0",
    "pytest>=8.4.1",
    "pytest-ansible>=25.6.3",
]

[tool.uv]
# Include dev dependencies by default for development workflow
default-groups = ["dev"]
# Manage the project environment
managed = true
# Enable package mode for CLI entry point installation
package = true
# Performance optimizations
concurrent-builds = 4
concurrent-downloads = 8
concurrent-installs = 4
# Better development experience
compile-bytecode = false      # Keep false for faster dev iterations
python-preference = "managed" # Prefer uv-managed Python installations
# Caching optimization - invalidate cache when key files change
cache-keys = [
    { file = "pyproject.toml" },
    { file = "uv.lock" },
    { file = "ansible.cfg" },
    { file = "configs/domains.yml" },
]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = ["*.lock", "*.log", "tmp/", "logs/", ".venv/", "__pycache__/"]

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = ["--strict-markers", "--disable-warnings", "--color=yes"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.basedpyright]
# Target Python 3.13 for type checking
pythonVersion = "3.13"
# Include all Python files in the scripts directory
include = ["scripts"]
# Exclude common build/cache directories
exclude = [
    "**/__pycache__",
    "**/.pytest_cache",
    "**/node_modules",
    ".venv",
    "tmp",
    "logs",
]
# Type checking configuration
typeCheckingMode = "basic"
# Use the virtual environment for imports
venvPath = "."
venv = ".venv"
# Reporting configuration
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
reportImportCycles = "error"
reportUninitializedInstanceVariable = "warning"
reportCallInDefaultInitializer = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryComparison = "warning"
# Disable some overly strict checks for infrastructure code
reportConstantRedefinition = "none"
reportIncompatibleMethodOverride = "warning"
reportIncompatibleVariableOverride = "warning"
reportUnusedImport = "warning"
reportUnusedVariable = "warning"

[tool.pymarkdown]
# Configuration for auto-generated and infrastructure documentation
# Higher line length limit for auto-generated terraform documentation
plugins.md013.line_length = 500
plugins.md013.code_blocks = false
plugins.md013.tables = false
# Allow all HTML elements for auto-generated documentation
plugins.md033.allowed_elements = "*"
# Configure emphasis usage
plugins.md036.punctuation = ".,;:!?。，；：！？"
# Allow duplicate headers in different sections
plugins.md024.siblings_only = true
# Disable problematic rules for infrastructure documentation
disable_rules = ["MD033", "MD036"]
